spec: split run.spec resolution into _selection, keep _spec pure grammar
