+++ docs/source/_spec.rst
@@ -14,9 +14,11 @@ Selectors carry a category-prefixed plugin path (``probes.<module>[.<Class>]``,
 ``tier:<N|name>``), or the explicit empty selection ``none`` / ``probes.none``
 (distinct from an unspecified spec, which defaults to ``probes.*``). A leading
 ``-`` excludes; ``tier:N`` is inclusive ("log level": tiers ``1..N``).
-``Spec.resolve()`` orchestrates probe and buff selection plus the filters; the
-single plugin-path resolution core, ``_resolve_plugin_paths``, is shared with
-the ``parse_plugin_spec`` adapter used for detectors.
+
+``garak/_spec.py`` covers the grammar: parsing and serialisation. Resolving a
+``Spec`` to concrete plugin names (against active/tier/tag state) is done by
+``garak._selection.resolve_spec``; the single plugin-path resolution core is
+shared with the ``parse_plugin_spec`` adapter used for detectors.
 
 See :doc:`configurable` for the user-facing grammar and examples.
