+++ docs/source/index.rst
@@ -69,6 +69,7 @@ Check out the :doc:`usage` section for further information, including :doc:`inst
    intents
    payloads
    _plugins
+   _selection
    _spec
    cas
    report
